requirements.txt
python-dotenv==1.0.1
python-telegram-bot[webhooks]==20.3
openai==0.28.0
aiohttp>=3.8.1
pandas>=2.0.0
